+++ e-commerce/src/app/shared/services/commercetoolsApi/commercetoolsapi.service.ts
@@ -3,7 +3,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable, mergeMap } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { authVisitorAPI, unauthVisitorAPI } from '../../../../environment';
-import { Address, AuthData, CartBase, CustomerDraft, CustomerInfo, PasswordChange, PersonalInfo } from './apitypes';
+import { Address, AuthData, CustomerDraft, CustomerInfo, PasswordChange, PersonalInfo } from './apitypes';
 import TokenStorageService from '../tokenStorage/tokenstorage.service';
 import * as actions from '../../../store/actions';
 import { AppState } from '../../../store/store';
@@ -33,45 +33,6 @@ export default class CommerceApiService {
       .set('Authorization', `Basic ${btoa(`${unauthVisitorAPI.ctpClientId}:${unauthVisitorAPI.ctpClientSecret}`)}`);
 
     return this.http.post<AuthData>(unAuthUrl, body.toString(), { headers });
-  }
-
-  createAnonymousCart(accessToken: string): Observable<CartBase> {
-    const apiUrl = `${unauthVisitorAPI.ctpApiUrl}/${unauthVisitorAPI.ctpProjectKey}/me/carts`;
-
-    const body = {
-      currency: 'USD',
-    };
-
-    const headers = new HttpHeaders()
-      .set('Content-Type', 'application/json')
-      .set('Authorization', `Bearer ${accessToken}`);
-
-    return this.http.post<CartBase>(apiUrl, body, { headers });
-  }
-
-  updateAnonymousCart(accessToken: string, idCart: string, version: number, productId: string): Observable<CartBase> {
-    const apiUrl = `${unauthVisitorAPI.ctpApiUrl}/${unauthVisitorAPI.ctpProjectKey}/me/carts/${idCart}`;
-
-    const body = {
-      version,
-      actions: [
-        {
-          action: 'setCountry',
-          country: 'US',
-        },
-        {
-          action: 'addLineItem',
-          productId,
-          quantity: 1,
-        },
-      ],
-    };
-
-    const headers = new HttpHeaders()
-      .set('Content-Type', 'application/json')
-      .set('Authorization', `Bearer ${accessToken}`);
-
-    return this.http.post<CartBase>(apiUrl, body, { headers });
   }
 
   registration(customerDraft: CustomerDraft, anonToken: string, anonymousId: string = ''): Observable<CustomerDraft> {
